link layout: add TU split-feasibility planner and safe function mover
build/wip/m_tu_split_plan.py classifies every first-party .c file against
upstream's per-object splat map: 0 interleaved, 53 clean-cut, 21 line-overlap.
It also quantifies the two real blockers per file - file-scope statics
referenced across a cut, and preamble declarations a moved span depends on.

build/wip/m_tu_move.py performs a move with automatic revert on COMPILE_ERROR,
lost MATCH, or a function pushed over its window. Empirical result: of 18
single-function moves, 17 reverted, confirming declaration migration - not
interleaving - is what blocks TU convergence. One move landed
(k_command func_001c1e20 -> k_dungeon).

Both scripts preserve original line endings; restored frTagTable.c and
bp_effect.c to LF after an earlier round-trip flipped them to CRLF.

+++ src/Kosaka/k_vpad.c
@@ -60,7 +60,7 @@ void* K_VPad_UpdateTask(KwlnTask* task)
     s32 animation;
     s32 rotated;
     s32 forceAnimation;
-    s32 cameraInput;
+    u16 cameraInput;
     work = (VPadWork*)task->workData;
     kwlnGetMainCamera();
     axis = D_00683D78;
